+++ CHANGELOG.md
@@ -88,6 +88,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `binary_to_integer` and `list_to_integer` do not raise anymore `overflow` error, they raise
 instead `badarg`.
 - Resources are now references instead of empty binaries.
+- Badarg error return from calling crypto:crypto_one_time with invalid arguments now matches OTP24+.
 
 ### Fixed